# docs/swagger.yaml
basePath: /
definitions:
...
  job.CompleteJobRequest:
    description: Complete job request
    properties:
      final_value:
        example: 280
        type: integer
    type: object
  job.CreateJobRequest:
    description: Create job request
    properties:
      address:
        example: ul. Krolewska 10/5
        type: string
      budget:
        example: 200
        type: integer
      building_type:
        example: apartment
        type: string
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      contact_method:
        example: phone
        type: string
      description:
        example: Kran cieknie od tygodnia
        type: string
      district_id:
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
      floor:
        example: 3
        type: integer
      has_elevator:
        example: true
        type: boolean
      photo_urls:
        items:
          type: string
        type: array
      preferred_date1:
        type: string
      preferred_date2:
        type: string
      preferred_time:
        example: morning
        type: string
      title:
        example: Cieknacy kran w kuchni
        type: string
      urgency:
        example: normal
        type: string
      wants_invoice:
        example: false
        type: boolean
    type: object
  job.JobListResponse:
    description: Paginated job list
    properties:
      has_more:
        example: true
        type: boolean
      jobs:
        items:
          $ref: '#/definitions/job.JobResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 50
        type: integer
    type: object
  job.JobResponse:
    description: Job information
    properties:
      address:
        example: ul. Krolewska 10/5
        type: string
      budget:
        example: 200
        type: integer
      building_type:
        example: apartment
        type: string
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      client_confirmed:
        type: boolean
      client_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      completed_at:
        type: string
      completed_by_id:
        type: string
      contact_method:
        example: phone
        type: string
      created_at:
        example: "2025-01-10T12:00:00Z"
        type: string
      description:
        example: Kran cieknie od tygodnia, potrzebna wymiana uszczelki
        type: string
      district_id:
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
      expires_at:
        type: string
      final_value:
        example: 280
        type: integer
      floor:
        example: 3
        type: integer
      has_elevator:
        example: true
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      photo_urls:
        items:
          type: string
        type: array
      preferred_date1:
        type: string
      preferred_date2:
        type: string
      preferred_time:
        example: morning
        type: string
      status:
        example: active
        type: string
      title:
        example: Cieknacy kran w kuchni
        type: string
      updated_at:
        example: "2025-01-10T12:00:00Z"
        type: string
      urgency:
        example: normal
        type: string
      wants_invoice:
        example: false
        type: boolean
    type: object
...
  response.Error:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      error:
        example: Invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  response.ValidationError:
    properties:
      error:
        example: Validation failed
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@fixapp.com
    name: API Support
  description: FixApp backend API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FixApp API
  version: "1.0"
paths:
  /admin/jobs:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of all jobs (admin only)
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by category
        in: query
        name: category_id
        type: string
      - description: Filter by district
        in: query
        name: district_id
        type: string
      - description: Filter by client
        in: query
        name: client_id
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.JobListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: List all jobs (admin)
      tags:
      - admin
...
  /jobs:
    get:
      consumes:
      - application/json
      description: Returns the authenticated user's jobs with optional status filter
      parameters:
      - description: Filter by status (draft, active, accepted, in_progress, done,
          cancelled)
        in: query
        name: status
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.JobListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: List my jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Creates a new service request (job) for the authenticated client
      parameters:
      - description: Job details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/job.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/job.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ValidationError'
      security:
      - BearerAuth: []
      summary: Create a new job
      tags:
      - jobs
  /jobs/{id}:
    get:
      consumes:
      - application/json
      description: Returns a job by its ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.JobResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get job by ID
      tags:
      - jobs
  /jobs/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancels a job (client can cancel their own, admin can cancel any)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Cancel a job
      tags:
      - jobs
  /jobs/{id}/complete:
    post:
      consumes:
      - application/json
      description: Marks a job as done with a declared final value (handyman only)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Completion details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/job.CompleteJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Complete a job
      tags:
      - jobs
  /jobs/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Client confirms that the job was completed successfully
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Confirm job completion
      tags:
      - jobs
  /jobs/{id}/publish:
    post:
      consumes:
      - application/json
      description: Moves a draft job to active status, making it visible to handymen
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Publish a job
      tags:
      - jobs
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"